move cmd functions to cli.cpp
this is so that they can be put in a cmake library and linked for testing
