+++ cusk/tests/CMakeLists.txt
@@ -33,7 +33,7 @@ macro(package_add_test TESTNAME)
 
     # link the Google test infrastructure, mocking library, and a default main function to
     # the test executable.  Remove g_test_main if writing your own main function.
-    target_link_libraries(${TESTNAME} gtest gmock gtest_main block io prep parents)
+    target_link_libraries(${TESTNAME} gtest gmock gtest_main block io prep parents cli)
     target_include_directories(${TESTNAME} PRIVATE ../include)
 
     # gtest_discover_tests replaces gtest_add_tests,
